add code
0619_sample_image파일 추가.
csv파일 2개를 하나로 합침.
그것을 불러와서 이미지 위에 출력.
