+++ 1.data/3.DB/db.csv
@@ -1,0 +1,122 @@
+id,x1,y1,x2,y2
+0,443,132,544,148
+1,94,153,123,169
+2,131,153,155,168
+3,162,153,188,169
+4,220,153,277,168
+5,155,173,287,194
+6,155,198,297,223
+7,362,153,546,171
+8,362,171,548,188
+9,362,189,548,205
+10,362,205,547,223
+11,79,224,106,241
+12,106,224,135,241
+13,136,224,164,240
+14,163,224,193,241
+15,194,224,222,242
+16,79,242,108,258
+17,107,242,135,258
+18,137,242,165,258
+19,164,241,195,258
+20,195,242,222,259
+21,259,225,388,257
+22,423,223,546,242
+23,423,242,547,258
+24,45,134,61,147
+25,100,134,118,146
+26,158,134,174,147
+27,212,134,231,147
+28,282,134,300,149
+29,322,135,362,148
+30,43,306,275,318
+31,43,323,276,339
+32,43,341,274,360
+33,43,360,275,374
+34,43,378,276,391
+35,43,397,274,405
+36,43,414,273,427
+37,43,433,269,445
+38,43,449,271,466
+39,43,467,271,487
+40,43,484,269,502
+41,43,502,266,518
+42,43,521,272,533
+43,270,306,307,322
+44,270,323,309,342
+45,270,341,310,359
+46,270,360,309,374
+47,270,378,305,388
+48,270,397,311,404
+49,270,414,305,428
+50,270,433,299,459
+51,270,449,311,465
+52,270,467,312,481
+53,270,484,314,497
+54,270,502,310,516
+55,270,521,313,530
+56,304,306,349,319
+57,304,323,345,336
+58,304,341,341,355
+59,304,360,347,374
+60,304,378,347,395
+61,304,397,352,416
+62,304,414,350,438
+63,304,433,349,463
+64,304,449,346,479
+65,304,467,346,497
+66,304,484,314,497
+67,304,502,345,515
+68,304,521,348,532
+69,343,306,383,322
+70,343,323,381,337
+71,343,341,381,356
+72,343,360,380,374
+73,343,378,377,394
+74,343,397,377,412
+75,343,414,378,426
+76,343,433,379,447
+77,343,449,377,466
+78,343,467,381,481
+79,343,484,380,498
+80,343,502,378,522
+81,343,521,385,536
+82,381,306,536,519
+83,228,540,248,551
+84,302,540,322,548
+85,43,557,273,571
+86,43,574,270,587
+87,43,592,263,609
+88,43,612,258,628
+89,43,629,268,645
+90,43,648,264,663
+91,43,666,271,679
+92,272,557,306,572
+93,272,574,306,586
+94,272,592,303,606
+95,272,612,311,625
+96,272,629,312,642
+97,272,648,306,663
+98,272,666,311,670
+99,306,557,348,571
+100,306,574,344,588
+101,306,592,337,608
+102,306,612,342,626
+103,306,629,340,650
+104,306,648,342,668
+105,306,666,348,685
+106,344,557,388,570
+107,344,574,386,587
+108,344,592,382,608
+109,344,612,380,628
+110,344,629,374,650
+111,344,648,372,671
+112,344,666,372,671
+113,382,541,535,683
+114,162,683,232,696
+115,142,718,323,732
+116,166,738,284,748
+117,274,738,325,752
+118,142,752,328,766
+119,142,772,331,781
+120,325,735,543,783
